moved polling func to own directory
