+++ backend-peer/src/lib/services/cloudagent.ts
@@ -137,7 +137,8 @@ export class CloudagentManager {
     if (!res.ok) {
       throw new ServiceUnavailable('Error accepting implicit invitation')
     }
-    const responseBody: ImplicitInvitationResponse = await res.json()
+    const responseBody: ImplicitInvitationResponse =
+      (await res.json()) as ImplicitInvitationResponse
     if (waitUntilCompleted) {
       return await connectionRecordPromise
     } else {
@@ -291,7 +292,10 @@ export class CloudagentManager {
     const res = await fetch(`${this.url_prefix}/basic-messages/${connectionId}`)
 
     const responseBody = await res.json()
-
+    if (res.status === 500) {
+      const responseBody = (await res.json()) as AriesFrameworkError
+      throw new InternalError(responseBody.message)
+    }
     if (!res.ok) {
       throw new ServiceUnavailable('Error fetching cloud agent')
     }
