+++ test/e2e/connection.test.ts
@@ -4,6 +4,7 @@ import { describe } from 'mocha'
 import { KeyType, ConnectionRecord } from '@aries-framework/core'
 
 import { getConfig } from './fixtures/config.js'
+import { pollGetConnections } from './utils/shared.js'
 
 describe('Connection', async function () {
   const config = getConfig()
@@ -138,20 +139,3 @@ describe('Connection', async function () {
     })
   })
 })
-
-const pollGetConnections = async (
-  baseUrl: string,
-  stopPollCondition: (connections: ConnectionRecord[]) => boolean,
-  delay: number = 1000
-) => {
-  const poll = async (): Promise<void> => {
-    const { body: connections } = await request(baseUrl).get('/connections')
-    if (connections.length > 0 && stopPollCondition(connections)) {
-      return
-    }
-    await new Promise((resolve) => setTimeout(resolve, delay))
-    await poll()
-  }
-
-  await poll()
-}
