+++ test/e2e/messaging.test.ts
@@ -2,8 +2,9 @@ import { expect } from 'chai'
 import request from 'supertest'
 import { describe } from 'mocha'
 import { KeyType, ConnectionRecord } from '@aries-framework/core'
-
 import { getConfig } from './fixtures/config.js'
+
+import { pollGetConnections } from './utils/shared.js'
 
 describe('Messaging', async function () {
   const config = getConfig()
@@ -151,20 +152,3 @@ describe('Messaging', async function () {
     })
   })
 })
-
-const pollGetConnections = async (
-  baseUrl: string,
-  stopPollCondition: (connections: ConnectionRecord[]) => boolean,
-  delay: number = 1000
-) => {
-  const poll = async (): Promise<void> => {
-    const { body: connections } = await request(baseUrl).get('/connections')
-    if (connections.length > 0 && stopPollCondition(connections)) {
-      return
-    }
-    await new Promise((resolve) => setTimeout(resolve, delay))
-    await poll()
-  }
-
-  await poll()
-}
